requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
openai==0.28.1
Pillow==9.5.0
python-dotenv==1.0.0
python-pptx==0.6.22
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
XlsxWriter==3.1.9
yarl==1.9.2
